+++ teacher.rb
@@ -1,0 +1,12 @@
+require_relative 'person'
+
+class Teacher < Person
+  def initialize(age, specialization, name='unknown', parent_permission=true)
+    super(age, name, parent_permission)
+    @specialization = specialization
+  end
+
+  def can_use_services? 
+    true
+  end
+end
